% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findj.R
\name{findj}
\alias{findj}
\title{Find J}
\usage{
findj(mat)
}
\arguments{
\item{mat}{input matrix of counts for search}
}
\value{
number of the row j
}
\description{
For a matrix with elements presented as v_jk,
searches for the smallest j where v_jk>=k
}
\author{
Mitra Ebrahimpoor

\email{m.ebrahimpoor@lumc.nl}
}
