% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCat.R
\name{getCat}
\alias{getCat}
\title{Calculate p-categories}
\usage{
getCat(mps, gCT, scale = c("col", "row"))
}
\arguments{
\item{mps}{Matrix of p-values, representing pairwise associations between two feature sets.}

\item{gCT}{Parameters of the global closed testing, which is the output of simesCT function.}

\item{scale}{Scale of the quantification, a character string. Possible choices are "col" and "row".}
}
\value{
Matrix of p-categories.
}
\description{
Calculates cumulative p-categories for a given matrix of p-values.
}
\seealso{
\link{simesCT}
}
