% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocean.R
\name{ocean}
\alias{ocean}
\title{OCEAN algorithm}
\usage{
ocean(om1, om2, gCT, scale = c("pair", "row", "col"), mps, nMax = 100)
}
\arguments{
\item{om1, om2}{Subsets of omics data where rows are the probes and columns are samples.
The rows of the two matrices should define the two-way feature set of interest.}

\item{gCT}{Parameters of the global closed testing provided as the output of simesCT function}

\item{scale}{Optional, will specify the scale of the quantification, a character string. Possible choices are "pair", "col" and "row".
If not provided, all scales are returned.}

\item{mps}{Optional, a sub-matrix of the matrix of pairwise associations. If provided, om1 and om2 are not required.
If not provided, matrix of pairwise associations will be derived from om1 and om2 based on Pearson's correlation.}

\item{nMax}{Maximum number of steps for branch and bound, if set to 1 branch and bound
is skipped even if the result is not exact.}
}
\value{
TDP is returned for the scales defined by scale.
}
\description{
Estimates TDP bounds in 3 scales for a given pathway.
In case the results is not exact, the function adopts branch and bound if nMax allows.
}
\author{
Mitra Ebrahimpoor

\email{m.ebrahimpoor@lumc.nl}
}
