% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairTD.R
\name{pairTD}
\alias{pairTD}
\title{pairwise true discoveries proportion}
\usage{
pairTD(p, n, gCT)
}
\arguments{
\item{p}{Matrix of pairwise associations, can be passed as a vector}

\item{n}{Number of pairs; may not be the same as number of pairs in p if a
threshold is used to remove large p-values}

\item{gCT}{Parameters of the global closed testing provided as the output of simesCT function}
}
\value{
Proportion of true discoveries out of n pairs
}
\description{
Calculates the PTD over pair; based on SEA algorithm
}
\seealso{
[simesCT()]
}
\author{
Mitra Ebrahimpoor

\email{m.ebrahimpoor@lumc.nl}
}
