% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairTDP.R
\name{pairTDP}
\alias{pairTDP}
\title{pairwise true discoveries proportion}
\usage{
pairTDP(mps, n, gCT)
}
\arguments{
\item{mps}{Matrix or vector of pairwise associations.}

\item{n}{Number of pairs; may not be the size of p if a threshold is used to remove large p-values.}

\item{gCT}{Parameters of the global closed testing, output of simesCT function.}
}
\value{
Proportion of true discoveries out of n pairs of features.
}
\description{
Calculates the TDP over pairs; based on SEA algorithm
}
\seealso{
\link[rSEA]{SEA}, \link{simesCT}
}
\author{
Mitra Ebrahimpoor

\email{m.ebrahimpoor@lumc.nl}
}
