% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBaB.R
\name{runbab}
\alias{runbab}
\title{Branch and bound algorithm implementation}
\usage{
runbab(sCat, ssh, ssb, nMax = 100)
}
\arguments{
\item{sCat}{Category matrix, output of getCat function}

\item{ssh}{current Heuristic as provided by SingleStep function}

\item{ssb}{current Bound as provided by SingleStep function}

\item{nMax}{Maximum number of steps for the algorithm, the algorithm may stop sooner if it converges.}
}
\value{
A list, including the heuristic, the bound, number of steps
taken and convergence status
}
\description{
Performs B&B when the bound are not exact
}
\seealso{
[simesCT()], [getCat()]
}
\author{
Mitra Ebrahimpoor

\email{m.ebrahimpoor@lumc.nl}
}
