% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simesCT.R
\name{simesCT}
\alias{simesCT}
\title{Calculates parameters of global closed testing based on Simes test}
\usage{
simesCT(om1, om2, mps, m, alpha = 0.05)
}
\arguments{
\item{om1, om2}{Two omics datasets where rows are probs and columns are samples.}

\item{mps, m}{Optional, pre-calculated matrix/vector of pairwise associations and the size.
To save time in calculation of parameters, one can apply a threshold such as the type I error to rewove larger p-values.
If a threshold is used size of matrix and m will not match. m should always be the size of the original matrix of associations.}

\item{alpha}{type I error rate}
}
\value{
Vector of integers: grand H value, concentration p-value, size of concentration set z,
size of the original pair-wise associations matrix and the type I error level used in calculations.
}
\description{
Calculates five parameters from closed testing with Simes local tests based on row data.
 Grand H: size of the largest set which is not rejected by closed testing and z: size of concentration set.
 These values are unique per dataset/alpha-level combination and do not depend on pathways. Calculation may
 be somewhat long depending on the size of datasets and PC configurations.
}
\references{
See more details in "Hommel's procedure in linear time" doi: 10.1002/bimj.201700316.
}
\author{
Mitra Ebrahimpoor

\email{m.ebrahimpoor@lumc.nl}
}
