% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleStep.R
\name{singleStep}
\alias{singleStep}
\title{Single step algorithm}
\usage{
singleStep(sCat, B)
}
\arguments{
\item{sCat}{Category matrix, output of getCat function}

\item{B}{Optional, to identify rows to be fixed (1) or removed (0) while splitting the search space.}
}
\value{
A list of two objects, the heuristic and the bound
}
\description{
Calculates a heuristic and an upper-bound for the number of FD based on the algorithm
introduced in paper
}
\seealso{
[simesCT()], [getCat()], [findj()]
}
\author{
Mitra Ebrahimpoor

\email{m.ebrahimpoor@lumc.nl}
}
