% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corPs.R
\name{corPs}
\alias{corPs}
\title{Calculate pairwise p-value}
\usage{
corPs(pm1, pm2, type = c("Mat", "Vec"), pthresh = 0.05)
}
\arguments{
\item{pm1, pm2}{Subsets of two omics data sets where rows are the features and columns are samples.
The rows of the two matrices would define the two-way feature set of interest.}

\item{type}{Two options are available. Mat: Calculate the correlation of subsets and return a
matrix; Vec: calculate the correlation matrix, subset by the given threshold and return a
vector of p-values.}

\item{pthresh}{Only relevant for type="Vec". The threshold by which the p-values are filtered (p>pthresh is removed).
Default value is 0.05.}
}
\value{
A data structure (an ff object) including pairwise p-values.
}
\description{
Calculates pairwise matrix of p-values based on Pearson's correlation test for two matrices.
To gain speed and manage RAM usage, the matrices are split into several smaller chunks.
}
