% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCat.R
\name{getCat}
\alias{getCat}
\title{Calculate p-categories}
\usage{
getCat(ps, gCT, scale = c("col", "row"))
}
\arguments{
\item{ps}{Matrix of p-values, representing association between two omics pathways}

\item{gCT}{Parameters of the global closed testing provided as the output of simesCT function}

\item{scale}{Scale of the quantification, a character string. Possible choices are "col" and "row".}
}
\value{
Matrix of p-categories
}
\description{
Calculates cumulative p-categories for a matrix of p-values
}
\seealso{
[simesCT()]
}
\author{
Mitra Ebrahimpoor

\email{m.ebrahimpoor@lumc.nl}
}
