% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleStep.R
\name{singleStep}
\alias{singleStep}
\title{Single step algorithm}
\usage{
singleStep(sCat, B)
}
\arguments{
\item{sCat}{p-categories matrix, output of getCat function.}

\item{B}{Optional, to identify rows to be fixed (1) or removed (0) while splitting the search space.}
}
\value{
A list of two objects, the heuristic and the lower bound for true number of discoveries
}
\description{
Calculates heuristic and upper-bound for the number of true discoveries based on
the Algorithm 1 introduced in paper.
}
\seealso{
\link{getCat}
}
\author{
Mitra Ebrahimpoor

\email{m.ebrahimpoor@lumc.nl}
}
